legacy/python-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pandas==2.3.2
numpy==2.3.3
gspread==6.2.1
google-auth==2.40.3
python-dotenv==1.1.1